+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Version 4.1.1 *(In Development)*
 
  * Fix: `ShareActionProvider` in the split action bar no longer fills the entire
    screen.
+ * Fix: Automatically correct `ColorDrawable` not respecting bounds when used as
+   a stacked background.
 
 
 Version 4.1.0 *(2012-05-17)*